requirements.txt
Flask>=2.0
Werkzeug
google-genai